Added user_id column to list model, validation on endpoints

+++ server/routes/task_routes.py
@@ -10,11 +10,15 @@ task = Blueprint('task', __name__)
 @task.route('/tasks/<task_id>', methods=['GET', 'DELETE'])
 @protected_route
 def task_root(task_id):
+    user_id = g.user.get('id')
     method = request.method
 
     if method == 'GET':
         requested_task = Tasks.query.filter_by(id=task_id).first()
         
+        if requested_task.user_id != user_id:
+            return jsonify(msg="You can't perform this action."), 403
+
         response = {
             'uid': requested_task.uid,
             'id': requested_task.id,
@@ -23,6 +27,10 @@ def task_root(task_id):
         return jsonify(response), 200
     elif method == 'DELETE':
         requested_task = Tasks.query.filter_by(id=task_id).first()
+
+        if requested_task.user_id != user_id:
+            return jsonify(msg="You can't perform this action."), 403
+
         db.session.delete(requested_task)
         db.session.commit()
